fix(ci): compile Windows Tooling lifecycle tests
